fixed:important bug of changing windows behavior
--all the immediate changes in the tasks model were deleted. Instead the
signals are triggered and the application informs the tasks model
with the update signals. This had to be done in order for the windows
to represent correctly the three states. IMPORTANT, when the user
sets a window to be in AllActivities the application makes it also to
be in AllDesktops