Makes the qml code standalone
This commit breaks all the old functionaly of the plasmoid that wasn't
already broken but it is a needed step into moving towards a pure qml
approch. It removes the src directory from the build process and moves
src/package to package. The widget will run without the c++ wrapper code
but has lost all the functionalty that it provides.