IMPORTANT commit: cleaning up the models structure, restore store file
--the new models merge continues. The purpose of this is to remove
almost all the logic from qml for Activities and Workreas.

--the 0.2.1 storing workareas file comes back. For 0.2.2 we will use
the same file but for 0.2.3 there will be a new one which is going to
be a clean xml file (there wiil be an automatic loading in order to not
break the users workareas settings)

--The choice is finally made. There will be 3classes in C++ to handle
all the logic for activities and workareas. One for Activities, one for
workareas, and one to handle the cooperation between them. All the hacks
for which signaling mechanisms are used are going to be in separate C++ classes
like the plugins for the ActivityManager class