fixes #61, preparing updatewallpaper to be moved
--update wallpaper will go to the workareas data engine. I found
that wallpapers for normal activities can be found with out issues
without using containment(). For the add new activity through
the data engine there is no problem. For the "clone activity" the
only thing that has to be made is to disable the updatewallpaper
mechanism... :) In the new activity the background is the same
with the old one from which was copied... :)