fixes syncing between desktops and workareas
--desktops follow the workareas max number,
I am thinking about a Passive and Active mode for
workareas, in order for the user to be able to choose
the behavior. Passive mode could be that VDs are in
higher value for consistency instead of workareas.
Even though a pager is now easy for someone implement
as the workareas data engine exists.