cleaned up a bit TaskPreviewDeleg
--one step at a time, I used containsMouse in TaskPreviewDeleg
instead of onEntered, onExited. Next Step is going to be to
use the new way of dragging in order to simplify more. And
final step, is going to be split TaskPreviewDeleg in three
classes, BasicTaskDeleg->TaskPreviewDeleg->DesktopTaskPreviewDeleg.
the first one is going to provide only basic elements.
the second is going to support previews and dragging.
the third is going to provide also scrolling support.